fix: querydsl total_spending_amount 기본 타입 expression 정의

+++ pennyway-domain/src/main/java/kr/co/pennyway/domain/domains/spending/repository/SpendingCustomRepositoryImpl.java
@@ -32,9 +32,9 @@ public class SpendingCustomRepositoryImpl implements SpendingCustomRepository {
         TotalSpendingAmount result = queryFactory.select(
                         Projections.constructor(
                                 TotalSpendingAmount.class,
-                                spending.spendAt.year(),
-                                spending.spendAt.month(),
-                                spending.amount.sum()
+                                spending.spendAt.year().intValue(),
+                                spending.spendAt.month().intValue(),
+                                spending.amount.sum().longValue()
                         )
                 ).from(user)
                 .leftJoin(spending).on(user.id.eq(spending.user.id))
